+++ src/Managers/ConfigurationManager.h
@@ -15,14 +15,14 @@ public:
   static ConfigurationManager& Instance();
 
   const std::vector<Vec2>& getAssetConfigs(const std::string& type) const;
-  const std::vector<size_t>& getEntityConfigs(const std::string& type) const;
+  const std::vector<int>& getEntityConfigs(const std::string& type) const;
 
   void setAssetConfigs(std::map<std::string, std::vector<Vec2>>& assetConfigMap);
-  void setEntityConfigs(std::map<std::string, std::vector<size_t>>& entityConfigMap);
+  void setEntityConfigs(std::map<std::string, std::vector<int>>& entityConfigMap);
 
 private:
   ConfigurationManager();
 
   std::map<std::string, std::vector<Vec2>> m_assetConfigMap;
-  std::map<std::string, std::vector<size_t>> m_entityConfigMap;
+  std::map<std::string, std::vector<int>> m_entityConfigMap;
 };
